+++ apps/express-trial/src/main.ts
@@ -1,6 +1,6 @@
 import express, { Request } from 'express';
 import cors from 'cors';
-import { JustaName, SubnameUpdateRequest } from '@justaname.id/sdk';
+import { JustaName } from '@justaname.id/sdk';
 import dotenv from 'dotenv'
 
 dotenv.config();
@@ -25,6 +25,7 @@ const chainId = parseInt(process.env.JUSTANAME_CHAIN_ID as string);
 const domain = process.env.JUSTANAME_DOMAIN as string;
 const origin = process.env.JUSTANAME_ORIGIN as string;
 const apiKey = process.env.JUSTANAME_API_KEY as string;
+const ensDomain = process.env.JUSTANAME_ENS_DOMAIN as string;
 
 if(!origin) {
   throw new Error('Origin is required');
@@ -44,6 +45,10 @@ if (!domain) {
 
 if (!apiKey) {
   throw new Error('API Key is required');
+}
+
+if (!ensDomain) {
+  throw new Error('ENS Domain is required');
 }
 
 let justaname: JustaName;
@@ -93,7 +98,7 @@ app.post('/api/subnames/add', async (req: Request<SubnameAdd>, res) => {
   try {
     const add = await justaname.subnames.addSubname({
         username: username,
-        ensDomain: domain,
+        ensDomain: ensDomain,
         chainId: chainId
       },
       {
@@ -111,71 +116,6 @@ app.post('/api/subnames/add', async (req: Request<SubnameAdd>, res) => {
   }
 });
 
-export interface SubnameUpdate extends Omit<SubnameUpdateRequest, 'ensDomain' | 'chainId'> {
-  address: string;
-  signature: string;
-  message: string;
-}
-
-app.post('/api/subnames/update', async (req: Request<SubnameUpdate>, res) => {
-
-  const username = req.body.username;
-  const addresses = req.body.addresses;
-  const text = req.body.text;
-  const contentHash = req.body.contentHash;
-  const signature = req.body.signature;
-  const address = req.body.address;
-  const message = req.body.message;
-
-  if(!username) {
-    res.status(400).send({ message: 'Username is required' });
-    return;
-  }
-
-  if(!addresses) {
-    res.status(400).send({ message: 'Addresses are required' });
-    return;
-  }
-
-  if(!text) {
-    res.status(400).send({ message: 'Text records are required' });
-    return;
-  }
-
-  if(!contentHash && contentHash !== '') {
-    res.status(400).send({ message: 'Content hash is required' });
-    return;
-  }
-
-  if(!req.body.address || !req.body.signature || !req.body.message) {
-    res.status(400).send({ message: 'Address, signature and message are required' });
-    return;
-  }
-
-
-  try {
-    const add = await justaname.subnames.updateSubname({
-        username: username,
-        ensDomain: domain,
-        chainId: chainId,
-      addresses,
-      text,
-      contentHash
-      },
-      {
-        xSignature: signature,
-        xAddress: address,
-        xMessage: message
-      });
-
-    res.status(201).send(add);
-    return;
-  }
-  catch (error) {
-    if(error instanceof Error)
-      res.status(500).send({ error: error.message });
-  }
-})
 
 app.get('/api', (req, res) => {
   res.send({ message: 'Welcome to JustaName Express!' });
@@ -186,7 +126,6 @@ const server = app.listen(port, async  () => {
   justaname = JustaName.init({
     apiKey: process.env.JUSTANAME_API_KEY as string
   });
-
 
   console.log(`Listening at http://localhost:${port}/api`);
 });