+++ examples/backend/express/src/main.ts
@@ -25,6 +25,7 @@ const chainId = parseInt(process.env.JUSTANAME_CHAIN_ID as string);
 const domain = process.env.JUSTANAME_DOMAIN as string;
 const origin = process.env.JUSTANAME_ORIGIN as string;
 const apiKey = process.env.JUSTANAME_API_KEY as string;
+const ensDomain = process.env.JUSTANAME_ENS_DOMAIN as string;
 
 if(!origin) {
   throw new Error('Origin is required');
@@ -87,7 +88,7 @@ app.post('/api/subnames/add', async (req: Request<SubnameAdd>, res) => {
   try {
     const add = await justaname.subnames.addSubname({
       username: username,
-      ensDomain: domain,
+      ensDomain: ensDomain,
       chainId: chainId
     },
     {