+++ examples/fullstack/nextjs/app/src/app/api/subnames/add/route.ts
@@ -3,8 +3,6 @@ import { ChainId } from '@justaname.id/sdk';
 import { getJustaNameInstance } from '../../../../justaname';
 import { NextRequest } from 'next/server';
 
-
-
 export  async function POST(
   req: NextRequest
 ) {
@@ -26,10 +24,10 @@ export  async function POST(
   if(!message) {
     return new Response('Message is required', { status: 400 });
   }
-  const justaname = await getJustaNameInstance();
+  const justaname = getJustaNameInstance();
 
   const chainId = parseInt(process.env.JUSTANAME_CHAIN_ID as string) as ChainId
-  const ensDomain = process.env.JUSTANAME_DOMAIN as string
+  const ensDomain = process.env.JUSTANAME_ENS_DOMAIN as string
   try {
     const subname = await justaname.subnames.addSubname(
       {